requirements.txt
pip==22.2.2
pytz~=2022.2.1
numpy~=1.21.6
setuptools==65.3.0
pandas~=1.3.5
python-dateutil~=2.8.2
requests~=2.28.1
tweepy~=4.10.1
oauthlib~=3.2.0
urllib3~=1.26.12
certifi~=2022.6.15
six~=1.16.0
idna~=3.3
python-dotenv~=0.20.0
